refac: move schema and table name to DeltaTableOptions
